backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
redis==5.0.1
celery==5.3.4
requests==2.31.0
httpx==0.25.2
python-dotenv==1.0.0
langchain==0.1.0
openai==1.3.8
google-api-python-client==2.110.0
googlemaps==4.10.0
stripe==7.8.0
email-validator==2.1.0
jinja2==3.1.2
aiofiles==23.2.1
Pillow==10.1.0
boto3==1.34.0
